Fixup MedditUserService interface

+++ src/main/java/com/endregaswarriors/meddit/services/MedditUserService.java
@@ -4,6 +4,7 @@ import com.endregaswarriors.meddit.models.Response;
 import com.endregaswarriors.meddit.models.api.LoginUser;
 import com.endregaswarriors.meddit.models.api.NewUser;
 import com.endregaswarriors.meddit.models.database.MedditUser;
+import io.swagger.models.auth.In;
 import org.springframework.stereotype.Service;
 
 import java.util.concurrent.CompletableFuture;
@@ -13,8 +14,8 @@ public interface MedditUserService {
 
     CompletableFuture<Response<MedditUser>> signUserIn(NewUser newUser);
     CompletableFuture<Response<MedditUser>> loginUser(LoginUser loginUser);
-    CompletableFuture<Response<Void>> joinSubreddit(Long thread_id, Integer user_id);
-    CompletableFuture<Response<Void>> leaveSubreddit(Long thread_id, Integer user_id);
+    CompletableFuture<Response<Void>> joinSubreddit(Integer thread_id, Integer user_id);
+    CompletableFuture<Response<Void>> leaveSubreddit(Integer thread_id, Integer user_id);
     CompletableFuture<Response<Void>> upvoteThread(Long thread_id, Integer user_id, Boolean upvote);
     CompletableFuture<Response<Void>> upvoteComment(Long comment_id, Integer user_id, Boolean upvote);
 }
